use commandlineargsmanager to push or remove ue4commandline.txt from device
