+++ SpatialGDK/Source/SpatialGDKEditor/Public/SpatialGDKEditorModule.h
@@ -4,6 +4,7 @@
 #include "Modules/ModuleManager.h"
 
 class FLBStrategyEditorExtensionManager;
+class FSpatialGDKEditorCommandLineArgsManager;
 
 class FSpatialGDKEditorModule : public ISpatialGDKEditorModule
 {
@@ -12,6 +13,7 @@ public:
 	FSpatialGDKEditorModule();
 
 	SPATIALGDKEDITOR_API FLBStrategyEditorExtensionManager& GetLBStrategyExtensionManager() { return *ExtensionManager; }
+	SPATIALGDKEDITOR_API FSpatialGDKEditorCommandLineArgsManager& GetCommandLineArgsManager() { return *CommandLineArgsManager; }
 
 	virtual void StartupModule() override;
 	virtual void ShutdownModule() override;
@@ -39,4 +41,6 @@ private:
 	bool HandleCloudLauncherSettingsSaved();
 
 	TUniquePtr<FLBStrategyEditorExtensionManager> ExtensionManager;
+
+	TUniquePtr<FSpatialGDKEditorCommandLineArgsManager> CommandLineArgsManager;
 };
